fix: replace some import paths after moving entity.ts
affects: @haulmont/jmix-front-generator
